fixed the windows fs issues
